+++ src/Inputfield/InputfieldAsmSelect.php
@@ -1,6 +1,7 @@
 <?php
 namespace SRAG\ILIAS\Plugins\MetaData\Inputfield;
 
+use ilNonEditableValueGUI;
 use SRAG\ILIAS\Plugins\MetaData\Field\Field;
 use SRAG\ILIAS\Plugins\MetaData\Field\MultiDropdownField;
 use SRAG\ILIAS\Plugins\MetaData\FormProperty\ilAsmSelectInputGUI;
@@ -24,17 +25,24 @@ class InputfieldAsmSelect extends BaseInputfield
     public function getILIASFormInputs(Record $record)
     {
         $options = $this->field->options();
-        $input = new ilAsmSelectInputGUI($this->field->getLabel($this->lang), $this->getPostVar($record));
-        if ($this->field->getDescription($this->lang)) {
-            $input->setInfo($this->field->getDescription($this->lang));
-        }
-        $input->setRequired($options->isRequired());
         $data = array();
         foreach ($this->field->getData() as $field_data) {
             $data[$field_data->getId()] = $field_data->getValue($this->lang);
         }
+        if ($options->isOnlyDisplay()) {
+            $input = new ilNonEditableValueGUI($this->field->getLabel($this->lang));
+            $input->setValue(nl2br(implode("\n", array_map(function ($value) use ($data) {
+                return $data[$value];
+            }, $record->getValue(), $this->field->getData())), false));
+        } else {
+        $input = new ilAsmSelectInputGUI($this->field->getLabel($this->lang), $this->getPostVar($record));
+        $input->setRequired($options->isRequired());
         $input->setOptions($data);
         $input->setValue($record->getValue());
+        }
+        if ($this->field->getDescription($this->lang)) {
+            $input->setInfo($this->field->getDescription($this->lang));
+        }
 
         return array($input);
     }