+++ src/Inputfield/InputfieldSelect.php
@@ -1,6 +1,7 @@
 <?php
 namespace SRAG\ILIAS\Plugins\MetaData\Inputfield;
 
+use ilNonEditableValueGUI;
 use SRAG\ILIAS\Plugins\MetaData\Field\DropdownField;
 use SRAG\ILIAS\Plugins\MetaData\Record\Record;
 
@@ -22,11 +23,6 @@ class InputfieldSelect extends BaseInputfield
     public function getILIASFormInputs(Record $record)
     {
         $field_options = $this->field->options();
-        $input = new \ilSelectInputGUI($this->field->getLabel($this->lang), $this->getPostVar($record));
-        if ($this->field->getDescription($this->lang)) {
-            $input->setInfo($this->field->getDescription($this->lang));
-        }
-        $input->setRequired($field_options->isRequired());
         $options = array();
         foreach ($this->field->getData() as $field_data) {
             $options[$field_data->getId()] = $field_data->getValue($this->lang);
@@ -34,8 +30,18 @@ class InputfieldSelect extends BaseInputfield
         if ($field_options->isPrependEmptyOption()) {
             $options = array('' => '') + $options;
         }
+        if ($field_options->isOnlyDisplay()) {
+            $input = new ilNonEditableValueGUI($this->field->getLabel($this->lang));
+            $input->setValue($options[$record->getValue()]);
+        } else {
+        $input = new \ilSelectInputGUI($this->field->getLabel($this->lang), $this->getPostVar($record));
+        $input->setRequired($field_options->isRequired());
         $input->setOptions($options);
         $input->setValue($record->getValue());
+        }
+        if ($this->field->getDescription($this->lang)) {
+            $input->setInfo($this->field->getDescription($this->lang));
+        }
 
         return array($input);
     }