+++ plugin.php
@@ -3,7 +3,7 @@
 require_once __DIR__ . "/vendor/srag/dic/src/PHPVersionChecker.php";
 
 $id = 'sragmetadata';
-$version = '0.0.7';
+$version = '1.0.0';
 $ilias_min_version = '5.3.0';
 $ilias_max_version = '5.4.999';
 $responsible_mail = 'info@studer-raimann.ch';