+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Only show in certain places config
 - User field type
 - Delete function
+- `AbstractMetadataGUI` and `MetadataService`
 - ILIAS 5.4 support
 - Drop ILIAS 5.1 and ILIAS 5.2 support
 - Implement uninstaller